broadband-modem-mbim: fix warnings with -Wtype-limits
This change is exclusively to make the compiler happy, don't even like
it myself...

  mm-broadband-modem-mbim.c: In function ‘signal_state_query_ready’:
  /usr/include/glib-2.0/glib/gmacros.h:811:63: error: comparison of unsigned expression < 0 is always false [-Werror=type-limits]
    811 | #define CLAMP(x, low, high)  (((x) > (high)) ? (high) : (((x) < (low)) ? (low) : (x)))
        |                                                               ^
  mm-broadband-modem-mbim.c:1680:19: note: in expansion of macro ‘CLAMP’
   1680 |         quality = CLAMP (rssi == 99 ? 0 : rssi, 0, 31) * 100 / 31;
        |                   ^~~~~

+++ src/mm-broadband-modem-mbim.c
@@ -1677,7 +1677,7 @@ signal_state_query_ready (MbimDevice *device,
         guint32 quality;
 
         /* Normalize the quality. 99 means unknown, we default it to 0 */
-        quality = CLAMP (rssi == 99 ? 0 : rssi, 0, 31) * 100 / 31;
+        quality = CLAMP (rssi == 99 ? 0 : (gint)rssi, 0, 31) * 100 / 31;
 
         g_task_return_int (task, quality);
     } else
